requirements.txt
streamlit==1.42.0
google-api-python-client==2.162.0
google-auth==2.38.0
openai==1.63.2
pandas==2.2.3
python-dotenv==1.0.1
pytest==8.3.4
